chore(design system): move test config directory
