--- jest.config.js
@@ -1,8 +1,0 @@
-
-module.exports = {
-  setupFilesAfterEnv: [
-    '<rootDir>/jest.setup.js',
-    '<rootDir>/setupTests.js',
-    '@testing-library/jest-dom',
-  ],
-};
