+++ packages/bleh-ui/.babelrc
@@ -1,0 +1,8 @@
+
+{
+  "presets": [
+    "@babel/preset-env",
+    "@babel/preset-react",
+    ["@babel/preset-typescript", { "isTSX": true, "allExtensions": true }]
+  ]
+}
