+++ packages/bleh-ui/jest.config.js
@@ -1,0 +1,5 @@
+export const setupFilesAfterEnv = [
+  '<rootDir>/jest.setup.js',
+  '<rootDir>/setupTests.js',
+  '@testing-library/jest-dom',
+];
