+++ packages/bleh-ui/package.json
@@ -1,12 +1,17 @@
 {
   "name": "@bleh-ui/react",
   "module": "./dist/index.js",
-  "type": "module",
   "types": "./dist/index.d.ts",
+  "type": "module",
   "sideEffects": false,
   "devDependencies": {
+    "@babel/preset-env": "^7.26.7",
+    "@babel/preset-react": "^7.26.3",
     "@chakra-ui/cli": "^3.2.3",
+    "@chakra-ui/react": "^3.2.3",
+    "@commercetools-uikit/i18n": "^19.22.1",
     "@pandacss/types": "^0.47.1",
+    "@preconstruct/cli": "catalog:tooling",
     "@storybook/addon-a11y": "^8.5.1",
     "@storybook/addon-essentials": "^8.5.1",
     "@storybook/addon-interactions": "^8.5.1",
@@ -14,18 +19,33 @@
     "@storybook/react": "^8.5.1",
     "@storybook/react-vite": "^8.5.1",
     "@storybook/test": "^8.5.1",
+    "@testing-library/jest-dom": "^6.6.3",
+    "@testing-library/react": "15.0.6",
+    "@testing-library/user-event": "^14.5.1",
+    "@types/jest": "^29.5.14",
     "@types/lodash": "^4.17.13",
     "@types/react": "^18.3.12",
     "@types/react-dom": "^18.3.1",
     "@vitejs/plugin-react": "^4.3.4",
+    "babel-jest": "^29.7.0",
+    "history": "^5.3.0",
+    "hygen": "catalog:tooling",
+    "jest": "29.7.0",
+    "jest-environment-jsdom": "29.7.0",
+    "jest-puppeteer": "8.0.6",
+    "jest-runner-eslint": "1.2.0",
+    "jest-runner-stylelint": "2.3.7",
+    "jest-silent-reporter": "0.6.0",
+    "jest-validate": "29.7.0",
+    "jest-watch-typeahead": "2.2.2",
+    "react-router-dom": "^7.1.3",
     "storybook": "^8.5.1",
     "storybook-dark-mode": "^4.0.2",
+    "ts-jest": "^29.2.5",
+    "typescript": "catalog:tooling",
     "vite": "^5.4.10",
     "vite-plugin-dts": "^4.5.0",
-    "vite-tsconfig-paths": "^5.1.3",
-    "@testing-library/react": "^14.1.2",
-    "@testing-library/jest-dom": "^6.1.5",
-    "@testing-library/user-event": "^14.5.1"
+    "vite-tsconfig-paths": "^5.1.3"
   },
   "peerDependencies": {
     "react": "^18.3.1",
@@ -51,6 +71,8 @@
     "dev": "vite build --watch",
     "build-theme-typings": "pnpm chakra typegen ./src/theme/index.ts",
     "storybook": "storybook dev -p 6006",
-    "build-storybook": "storybook build"
+    "build-storybook": "storybook build",
+    "test": "jest --projects jest.{ts-test,test}.config.cjs",
+    "test:watch": "pnpm run test --watch"
   }
 }
