--- setupTests.js
@@ -1,5 +1,0 @@
-// import "@testing-library/jest-dom";
-const { TextEncoder, TextDecoder } = require('util');
-
-global.TextEncoder = TextEncoder;
-global.TextDecoder = TextDecoder;
